Fix login validation and add error handling in login routes

+++ routes/alumni.js
@@ -47,16 +47,19 @@ router.post("/register", async (req, res) => {
 
 router.post("/login", async (req, res) => {
     try {
-        const user = await Alumni.findOne({ username: req.body.username })
-        if (!user) return res.status(401).json({ message: "Invalid Username" })
+        if (("username" in req.body) && ("password" in req.body)) {
 
-        const result = await bcrypt.compare(req.body.password, user.password);
-        if (result) {
-            const token = jwt.sign(_.omit(user.toObject(), ["password"]), config.get("jwtPrivateKey"));
-            res.header("x-auth-token", token);
-            res.status(200).json(_.omit(user.toObject(), ["password", "__v"]));
-        }
-        else res.status(401).json({ message: "Invalid Password" })
+            const user = await Alumni.findOne({ username: req.body.username })
+            if (!user) return res.status(401).json({ message: "Invalid Username" })
+
+            const result = await bcrypt.compare(req.body.password, user.password);
+            if (result) {
+                const token = jwt.sign(_.omit(user.toObject(), ["password"]), config.get("jwtPrivateKey"));
+                res.header("x-auth-token", token);
+                res.status(200).json(_.omit(user.toObject(), ["password", "__v"]));
+            }
+            else res.status(401).json({ message: "Invalid Password" })
+        } else return res.status(403).json({ message: "Username and password are two required fields" })
     }
     catch (err) {
         console.log(err);
@@ -76,14 +79,19 @@ router.put("/update/:id", alumniAuth, validateObjectId, async (req, res) => {
         }
         const user = await Alumni.findById(req.params.id);
         if (user) {
-            const { error } = validateProps(req.body);
-            if (error) return res.status(400).json({ message: error.details[0].message });
-
             if (req.username === user.username) {
+
+                const { error } = validateProps(req.body);
+                if (error) return res.status(400).json({ message: error.details[0].message });
+
                 user.set(req.body);
                 const result = await user.save();
                 res.status(200).json(_.omit(result.toObject(), ["password", "__v"]));
             } else return res.status(403).json({ message: "You cannot change other user data by providing their id" })
+
+            const { error } = validateProps(req.body);
+            if (error) return res.status(400).json({ message: error.details[0].message });
+
         } else {
             res.status(404).json({ message: "No user found with that alumniId" });
         }
@@ -98,12 +106,15 @@ router.put("/update/:id", alumniAuth, validateObjectId, async (req, res) => {
 
 router.delete("/delete/:id", alumniAuth, validateObjectId, async (req, res) => {
     try {
-        const result = await Alumni.findByIdAndDelete(req.params.id);
-        if (result) {
-            res.status(204).json({ message: "User deleted successfully" });
-        } else {
-            res.status(404).json({ message: "No user found with that alumniId" });
-        }
+        const user = await Alumni.findById(req.params.id);
+        if (req.username === user.username) {
+            const result = await Alumni.findByIdAndDelete(req.params.id);
+            if (result) {
+                res.status(204).json({ message: "User deleted successfully" });
+            } else {
+                res.status(404).json({ message: "No user found with that alumniId" });
+            }
+        } else res.status(403).json({ message: "you cannot delete other user data by providing their id" })
     } catch (err) {
         console.log(err);
         res.status(500).json({ message: "An error occurred while deleting the user" });
